Merge in RIV/rival-hub-core from RIV-41-dodawanie-stanowiska to develop

* commit 'd28b2a251c98fdd200e8468638ce67256657ace9':
  RIV-41 dodawanie stanowiska

+++ src/main/java/com/rivalhub/organization/OrganizationRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OrganizationRepository extends PagingAndSortingRepository<Organization, Integer>, CrudRepository<Organization, Integer> {
+public interface OrganizationRepository extends PagingAndSortingRepository<Organization, Long>, CrudRepository<Organization, Long> {
 
 }
